+++ Plugin/src/Content/Weathers/GodRayWeather.cs
@@ -84,18 +84,12 @@ public class GodRay(Color skyColour, Vector2 topPosition, float topRadius, float
 
 public class GodRayManager : CodeRebirthWeathers
 {
-    // these lists must be kept aligned
-    // so that godRayEffects[i] os the effect for godRays[i],
-    // for example.
-    // you might see the number of lists and call it "a bad idea",
-    // but I call it "ECS", which sounds much fancier.
-    List<GodRaySkyEffect> godRayEffects = [];
-    List<HDAdditionalLightData> godRaySpotlights = [];
-    List<GodRay> godRays = [];
+    List<GodRaySkyEffect> godRayEffects = new();
+    List<HDAdditionalLightData> godRaySpotlights = new();
+    List<GodRay> godRays = new();
 
     public Material godRayMaterial = null!;
     public GameObject godRayParent = null!;
-
     public Material sphereMaterial = null!;
     ComputeBuffer? rayBuffer;
     Camera localCamera = null!;
@@ -105,15 +99,14 @@ public class GodRayManager : CodeRebirthWeathers
     [SerializeField]
     public List<Color> rayColours;
     System.Random godRayRandom = new();
-	public static GodRayManager? Instance { get; private set; }
-	public static bool Active => Instance != null;
+    public static GodRayManager? Instance { get; private set; }
+    public static bool Active => Instance != null;
+
     private void OnEnable()
     {
         Instance = this;
         timeBetweenGodRaySpawns = 10f;
-        rayColours = [
-            new(1f, 0f, 0f, 0.25f),
-        ];
+        rayColours = new List<Color> { new Color(1f, 0f, 0f, 0.25f) };
         godRayRandom = new System.Random(StartOfRound.Instance.randomMapSeed);
         localCamera = GameNetworkManager.Instance.localPlayerController.gameplayCamera;
         transform.position = localCamera.transform.position;
@@ -123,28 +116,24 @@ public class GodRayManager : CodeRebirthWeathers
         StartCoroutine(UpdateGodRays());
     }
 
-    private IEnumerator UpdateGodRays() {
-        // Calculate the center point from all AI nodes
+    private IEnumerator UpdateGodRays()
+    {
         Vector3 centre = CalculateCenterOfPoints(RoundManager.Instance.outsideAINodes.Select(node => node.transform.position).ToList());
 
-        while (this.godRays.Count() <= 10) {
+        while (godRays.Count() <= 10)
+        {
             yield return new WaitForSeconds(godRayRandom.NextFloat(0.75f, 1.25f) * timeBetweenGodRaySpawns);
-            
-            // Get a random bottom position within a certain range around the center
+
             Vector3 bottomPosition = RoundManager.Instance.GetRandomNavMeshPositionInBoxPredictable(centre, 100f, default, godRayRandom);
-            
-            // Top position can be directly above the bottom position
-            Vector2 position = this.godRays.Count() == 0 ? Vector2.zero : new Vector2(bottomPosition.x, bottomPosition.z);
+            Vector2 position = godRays.Count() == 0 ? Vector2.zero : new Vector2(bottomPosition.x, bottomPosition.z);
             Color rayColour = rayColours[godRayRandom.Next(0, rayColours.Count)];
 
-            // Create and add the GodRay
-            this.AddGodRay(new GodRay(rayColour, position, godRayRandom.NextFloat(2f, 4f), godRayRandom.NextFloat(2f, 5f), new Vector3(bottomPosition.x, 0, bottomPosition.z), godRayRandom.NextFloat(3f, 8f), rayColour));
+            AddGodRay(new GodRay(rayColour, position, godRayRandom.NextFloat(2f, 4f), godRayRandom.NextFloat(2f, 5f), new Vector3(bottomPosition.x, 0, bottomPosition.z), godRayRandom.NextFloat(3f, 8f), rayColour));
         }
     }
 
     private void SetScale(float scale)
     {
-        // just to prevent the furthest vertex from being clipped
         scale *= 0.99f;
         transform.localScale = Vector3.one * scale * 2;
         transform.position = localCamera.transform.position;
@@ -194,9 +183,9 @@ public class GodRayManager : CodeRebirthWeathers
     /// <returns>true if the ray has been removed, or false otherwise</returns>
     private bool RemoveGodRay(GodRay ray)
     {
-        Plugin.ExtendedLogging($"Adding GodRay at TopPosition: {ray.TopPosition}, BottomPosition: {ray.BottomPosition}");
-        int index;
-        if ((index = godRays.IndexOf(ray)) != -1)
+        Plugin.ExtendedLogging($"Removing GodRay at TopPosition: {ray.TopPosition}, BottomPosition: {ray.BottomPosition}");
+        int index = godRays.IndexOf(ray);
+        if (index != -1)
         {
             godRays.RemoveAt(index);
             godRayEffects.RemoveAt(index);
@@ -207,7 +196,6 @@ public class GodRayManager : CodeRebirthWeathers
         }
         return false;
     }
-
     /// <summary>
     /// Spawns a GodRay.
     /// </summary>
@@ -215,7 +203,6 @@ public class GodRayManager : CodeRebirthWeathers
     /// <returns>The ray that was added (a reference to the given argument "ray")</returns>
     private GodRay AddGodRay(GodRay ray)
     {
-        Plugin.ExtendedLogging($"Adding GodRay at TopPosition: {ray.TopPosition}, BottomPosition: {ray.BottomPosition}");
         float skyHeight = localCamera.farClipPlane * 0.99f + localCamera.transform.position.y;
         godRays.Add(ray);
         godRayEffects.Add(ray.SkyEffect(skyHeight));
@@ -224,7 +211,7 @@ public class GodRayManager : CodeRebirthWeathers
         HDAdditionalLightData light = lightGameObject.AddHDLight(HDLightTypeAndShape.ConeSpot);
         light.range = localCamera.farClipPlane * 2;
 
-        float innerAnglePercent = Mathf.Clamp((spotlightData.angle*1.5f) * Mathf.Rad2Deg, 1f, 179f);
+        float innerAnglePercent = Mathf.Clamp((spotlightData.angle * 1.5f) * Mathf.Rad2Deg, 1f, 179f);
         if (spotlightData.angle * Mathf.Rad2Deg < 1f) innerAnglePercent = spotlightData.angle * Mathf.Rad2Deg * 100f;
         light.SetSpotAngle(innerAnglePercent, innerAnglePercent);
         light.shapeRadius = 0;
@@ -290,7 +277,7 @@ public class GodRayManager : CodeRebirthWeathers
 
     private void RegenerateRayComputeBuffer()
     {
-        if (godRayEffects.Count != 0)
+        if (godRayEffects.Count > 0)
         {
             if (rayBuffer == null || rayBuffer.count != godRayEffects.Count)
             {
@@ -308,37 +295,13 @@ public class GodRayManager : CodeRebirthWeathers
         sphereMaterial.SetInt("_rayCount", godRayEffects.Count);
     }
 
-    private IEnumerator FadeInLight(HDAdditionalLightData light, float targetIntensity, float duration)
-    {
-        float initialIntensity = light.intensity;
-        for (float t = 0; t < duration; t += Time.deltaTime)
-        {
-            light.SetIntensity(Mathf.Lerp(initialIntensity, targetIntensity, t / duration), LightUnit.Lux);
-            yield return null;
-        }
-        light.SetIntensity(targetIntensity, LightUnit.Lux);
-    }
-
-    private IEnumerator FadeOutLight(HDAdditionalLightData light, float duration)
-    {
-        float initialIntensity = light.intensity;
-        for (float t = 0; t < duration; t += Time.deltaTime)
-        {
-            light.SetIntensity(Mathf.Lerp(initialIntensity, 0, t / duration), LightUnit.Lux);
-            yield return null;
-        }
-        light.SetIntensity(0, LightUnit.Lux);
-        light.gameObject.SetActive(false); // Optionally return to pool here
-    }
-
     private void OnDisable()
     {
         Instance = null;
-
         StopAllCoroutines();
         foreach (GodRay ray in godRays.ToList())
         {
-            this.RemoveGodRay(ray);
+            RemoveGodRay(ray);
         }
         godRays.Clear();
         godRayEffects.Clear();